+++ package/src/components/index.ts
@@ -14,7 +14,7 @@ export * from './AttachmentPicker/AttachmentPicker';
 export * from './AttachmentPicker/components/AttachmentPickerBottomSheetHandle';
 export * from './AttachmentPicker/components/AttachmentPickerError';
 export * from './AttachmentPicker/components/AttachmentPickerErrorImage';
-export * from './AttachmentPicker/components/AttachmentSelectionBar';
+export * from './AttachmentPicker/components/AttachmentPickerSelectionBar';
 export * from './AttachmentPicker/components/CameraSelectorIcon';
 export * from './AttachmentPicker/components/FileSelectorIcon';
 export * from './AttachmentPicker/components/ImageOverlaySelectedComponent';