+++ package/src/contexts/overlayContext/OverlayContext.tsx
@@ -24,7 +24,6 @@ export type OverlayContextValue = {
   overlay: Overlay;
   setOverlay: React.Dispatch<React.SetStateAction<Overlay>>;
   style?: DeepPartial<Theme>;
-  translucentStatusBar?: boolean;
 };
 
 export const OverlayContext = React.createContext(
@@ -37,7 +36,10 @@ export type OverlayProviderProps<
   Partial<
     Pick<
       AttachmentPickerContextValue,
+      | 'AttachmentPickerBottomSheetHandle'
+      | 'attachmentPickerBottomSheetHandleHeight'
       | 'attachmentPickerBottomSheetHeight'
+      | 'AttachmentPickerSelectionBar'
       | 'attachmentSelectionBarHeight'
       | 'bottomInset'
       | 'CameraSelectorIcon'
@@ -56,8 +58,7 @@ export type OverlayProviderProps<
       | 'OverlayReactions'
       | 'OverlayReactionsAvatar'
     >
-  > &
-  Pick<OverlayContextValue, 'translucentStatusBar'> & {
+  > & {
     autoPlayVideo?: boolean;
     /**
      * The giphy version to render - check the keys of the [Image Object](https://developers.giphy.com/docs/api/schema#image-object) for possible values. Uses 'fixed_height' by default