+++ package/src/contexts/themeContext/utils/theme.ts
@@ -476,10 +476,13 @@ export type Theme = {
       shuffle: TextStyle;
       title: TextStyle;
     };
+    lastMessageContainer: ViewStyle;
     loadingIndicator: {
       container: ViewStyle;
       roundedView: ViewStyle;
     };
+    messageGroupedSingleOrBottomContainer: ViewStyle;
+    messageGroupedTopContainer: ViewStyle;
     pinnedHeader: {
       container: ViewStyle;
       label: TextStyle;
@@ -970,10 +973,7 @@ export const defaultTheme: Theme = {
         fontWeight: '700',
         paddingRight: 6,
       },
-      metaContainer: {
-        flexDirection: 'row',
-        marginTop: 4,
-      },
+      metaContainer: {},
       metaText: {
         fontSize: 12,
       },
@@ -1025,10 +1025,13 @@ export const defaultTheme: Theme = {
       shuffle: {},
       title: {},
     },
+    lastMessageContainer: {},
     loadingIndicator: {
       container: {},
       roundedView: {},
     },
+    messageGroupedSingleOrBottomContainer: {},
+    messageGroupedTopContainer: {},
     pinnedHeader: {
       container: {},
       label: {},